+++ src/utils.js
@@ -40,7 +40,7 @@ const utils = {
         addDeity: true,
         addName: true,
         addClass: true,
-        addPets: true,
+        addFamiliars: true,
         addFormulas: true,
         askForChoices: false,
       };
@@ -56,6 +56,46 @@ const utils = {
 
   resetFlags: async (actor) => {
     return utils.setFlags(actor, null);
+  },
+
+
+  getOrCreateFolder: async (root, entityType, folderName, folderColor = "") => {
+    let folder = game.folders.contents.find((f) =>
+      f.type === entityType && f.name === folderName
+      // if a root folder we want to match the root id for the parent folder
+      && (root ? root.id : null) === (f.folder?.id ?? null)
+    );
+    // console.warn(`Looking for ${root} ${entityType} ${folderName}`);
+    // console.warn(folder);
+    if (folder) return folder;
+    folder = await Folder.create(
+      {
+        name: folderName,
+        type: entityType,
+        color: folderColor,
+        parent: (root) ? root.id : null,
+      },
+      { displaySheet: false }
+    );
+    return folder;
+  },
+
+  // eslint-disable-next-line no-unused-vars
+  getFolder: async (kind, subFolder = "", baseFolderName = "Pathmuncher", baseColor = "#6f0006", subColor = "#98020a", typeFolder = true) => {
+    let entityTypes = new Map();
+    entityTypes.set("pets", "Pets");
+
+    const folderName = game.i18n.localize(`${CONSTANTS.MODULE_NAME}.labels.${kind}`);
+    const entityType = entityTypes.get(kind);
+    const baseFolder = await utils.getOrCreateFolder(null, entityType, baseFolderName, baseColor);
+    const entityFolder = typeFolder ? await utils.getOrCreateFolder(baseFolder, entityType, folderName, subColor) : baseFolder;
+    if (subFolder !== "") {
+      const subFolderName = subFolder.charAt(0).toUpperCase() + subFolder.slice(1);
+      const typeFolder = await utils.getOrCreateFolder(entityFolder, entityType, subFolderName, subColor);
+      return typeFolder;
+    } else {
+      return entityFolder;
+    }
   },
 
 };